+++ .gitignore
@@ -397,3 +397,4 @@ FodyWeavers.xsd
 
 # JetBrains Rider
 *.sln.iml
+/Graeae.Models/OpenApi.Models.xml
